FIX: Comprehensive mixed-precision dtype handling (v2)
Based on deep research into PyTorch dtype behavior:

1. COLAB_TRITON_TEST.py:
   - Remove .squeeze(2) bug - spectral_value_forward expects 4D input

2. spectral_cache.py:
   - Convert coeffs_q to model dtype after INT8 simulation
   - Fixed both _compress_tensor and _compress_tensor_fallback paths
   - SVD returns FP32, downstream ops need matching dtype

3. spectral_attention.py:
   - Define model_dtype at function start for consistent output
   - Use FP32 for einsum (NOT on autocast whitelist)
   - Convert scores back to model_dtype after computation
   - Apply same fix to value aggregation (attn @ coeffs @ basis)

4. kernels/spectral_attention.py:
   - spectral_score_forward_pytorch: FP32 einsum, convert back
   - spectral_value_forward_pytorch: FP32 einsum, convert back
   - Added dtype logging for debugging

Research findings applied:
- einsum NOT on autocast whitelist, requires explicit FP32 cast
- SVD always returns FP32, must convert ALL components back
- Mixed FP32/FP16 in einsum causes 'expected Float but found Half'
- BF16 is typical compute_dtype for Unsloth on modern GPUs

+++ unsloth_spectral/spectral_cache.py
@@ -306,7 +306,11 @@ class SpectralCache:
         # Batched quantization (vectorized per-head)
         coeffs_q, scales = self._quantize_int8_batched(coeffs)
         
-        # Keep basis in FP16 (it's small, no need to quantize)
+        # CRITICAL: Convert ALL spectral components back to model dtype (BF16/FP16)
+        # SVD computes in FP32 for numerical stability, but downstream ops need matching dtypes.
+        # Research finding: einsum is NOT on autocast whitelist, requires explicit dtype match.
+        # Without this, mixed FP32 coeffs × FP16 queries cause "expected Float but found Half".
+        coeffs_q = coeffs_q.to(self.dtype)
         basis = basis.to(self.dtype)
         
         return coeffs_q, basis, scales
@@ -343,7 +347,8 @@ class SpectralCache:
             coeffs_h = U_k * S_k.unsqueeze(0)
             coeffs_q, scale = self._quantize_int8(coeffs_h)
             
-            coeffs_list.append(coeffs_q)
+            # Convert coeffs to model dtype (matching basis)
+            coeffs_list.append(coeffs_q.to(self.dtype))
             basis_list.append(Vh_k.to(self.dtype))
             scales_list.append(scale)
